backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
python-dotenv==1.0.0
pydantic==1.10.13
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0
sqlalchemy-utils==0.41.1
